pyproject.toml
[tool.poetry]
name = "morpher"
version = "0.1.0"
description = "Dynamic agent that can modify the plan of execution on the fly."
authors = ["Snehal Reddy Yelmati <snehalyelmati@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.6.1"
langchain = "^0.0.352"
pydantic = "^2.5.3"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
jupyterlab = "^4.0.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
